Add dio dependency & refactor
