+++ lib/util/app_init.dart
@@ -1,4 +1,4 @@
-import 'package:places/utils/consts.dart';
+import 'package:places/util/consts.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// Класс для инициализации приложения
