+++ pubspec.yaml
@@ -13,6 +13,8 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
+  #  A powerful Http client for Dart
+  dio: ^3.0.10
   # Widget that automatically resizes text to fit perfectly within its bounds
   auto_size_text: ^3.0.0-nullsafety.0
   # SVG image
